Moving base image from custom CentOS to upstream python for bug 1431275
This is mostly moving from CentOS to debian. I've went for stretch,
because Jessie comes with really old versions of git and hg, and
that's bad for all kinds of things.

I didn't manage to figure out how to install packages on -slim, so
I'm using the full base images now.