hg poller should wait for empty queue on start-up
We're splitting out the hg poller from the a10n hg worker, so
the poller might be running while the worker is down. And then
restart, and then it tries to submit things to the queue again.

Don't do that, instead wait for the queue to be empty before
adding more stuff.

.message_count isn't reliable, but we don't need it to be. We're
the only one writing, and we know that we don't want to write
until it's empty, aka, message_count is just 0.